Miscellaneous cleanup, moved pairwise to utils.py
